[Web] Consolidate data sources into PaginatedDataSource class
