+++ attendance-web/src/app/components/students/show-student/show-student.component.ts
@@ -1,7 +1,5 @@
-import { CollectionViewer, DataSource } from '@angular/cdk/collections';
 import { HttpErrorResponse } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
-import { PageEvent } from '@angular/material/paginator';
 import { ActivatedRoute } from '@angular/router';
 import { DateTime } from 'luxon';
 import { AsyncSubject, BehaviorSubject, catchError, filter, map, Observable, of, ReplaySubject, share, switchMap, tap, throwError } from 'rxjs';
@@ -12,6 +10,7 @@ import { AttendanceService } from 'src/app/services/attendance.service';
 import { ErrorService } from 'src/app/services/error.service';
 import { StudentsService } from 'src/app/services/students.service';
 import { UsersService } from 'src/app/services/users.service';
+import { PaginatedDataSource } from 'src/app/utils/PaginatedDataSource';
 
 enum PageState {
   STUDENT_LOADING = 1,
@@ -64,40 +63,6 @@ class AttendanceStats {
   }
 }
 
-class SessionDataSource implements DataSource<RichAttendanceSession> {
-  private data: Array<RichAttendanceSession> = [];
-
-  public readonly pageSizeOptions = [25, 50, 100];
-
-  private paginatedData = new ReplaySubject<Array<RichAttendanceSession>>(1);
-
-  private lastPageSize = this.pageSizeOptions[0];
-
-  public setData(data: Array<RichAttendanceSession>): void {
-    this.data = data;
-    this.updateSlice(0, this.lastPageSize);
-  }
-
-  public paginate(event: PageEvent) {
-    this.lastPageSize = event.pageSize;
-    this.updateSlice(event.pageSize * event.pageIndex, event.pageSize * (event.pageIndex + 1));
-  }
-
-  public getDataSize(): number {
-    return this.data.length;
-  }
-
-  private updateSlice(startIdx: number, endIdx: number) {
-    this.paginatedData.next(this.data.slice(startIdx, endIdx));
-  }
-
-  connect(collectionViewer: CollectionViewer): Observable<readonly RichAttendanceSession[]> {
-    return this.paginatedData;
-  }
-
-  disconnect(collectionViewer: CollectionViewer): void {}
-}
-
 @Component({
   selector: 'app-show-student',
   templateUrl: './show-student.component.html',
@@ -111,7 +76,7 @@ export class ShowStudentComponent implements OnInit {
 
   protected student = new AsyncSubject<Student>()
   protected registeredBy = new AsyncSubject<User>()
-  protected attendanceSessions = new SessionDataSource()
+  protected attendanceSessions = new PaginatedDataSource<RichAttendanceSession>()
   protected attendanceStats = new AsyncSubject<AttendanceStats>()
 
   protected sessionColumns = ["checkInDate", "checkOutDate", "duration"]
